requirements.txt
streamlit
pandas
PIL
seaborn
matplotlib.pyplot